feat: Add deleteItemFromCart functionality to cartController
The code changes in main.js add a new function `deleteItemFromCart` to the `cartController` in order to delete an item from the cart. This function sends a DELETE request to the server with the item ID and updates the cart view accordingly. This allows users to easily remove items from their cart.

This commit message follows the convention of starting with a type (`feat` for feature) and providing a concise and descriptive summary of the changes made.