feat: Add finishCart method to CartController and CartRepository
The code changes in CartController and CartRepository add a new method called finishCart. This method allows the user to finish a cart by updating the 'closed' field in the database. The finishCart method is called when the user clicks the "Finalizar Compra" button in the cart page. This commit enhances the functionality of the cart feature by allowing users to complete their purchases.